Fix faucet link (#73)
Fixes https://github.com/paritytech/contract-issues/issues/74

The faucet is about to be changed to only reference the network and not
individual chains. Tokens will always be sent to AssetHub.

I will remove all references to AssetHub from the docs in a follow up.
It will only be references as Westend.